1_1_5 fix double alarm message, change list to eye

+++ app/src/main/java/ru/wintrade/mvp/presenter/traderme/TraderMeObservationPresenter.kt
@@ -31,6 +31,7 @@ class TraderMeObservationPresenter : MvpPresenter<TraderMeObservationView>() {
     lateinit var resourceProvider: ResourceProvider
 
     var listPresenter = TraderMeObservationListPresenter()
+    var subscriptionClickPos = -1
 
     inner class TraderMeObservationListPresenter : IObservationListPresenter {
         var traders = mutableListOf<Subscription>()
@@ -79,21 +80,44 @@ class TraderMeObservationPresenter : MvpPresenter<TraderMeObservationView>() {
             if (profile.user == null) {
                 router.navigateTo(Screens.signInScreen(false))
             } else {
-                apiRepo
-                    .deleteObservation(profile.token!!, traders[pos].trader.id)
-                    .observeOn(AndroidSchedulers.mainThread())
-                    .subscribe({}, {})
-                reLoadSubscriptions()
+                if (needDeleteSubscription(pos)) {
+                    apiRepo
+                        .deleteObservation(profile.token!!, traders[pos].trader.id)
+                        .observeOn(AndroidSchedulers.mainThread())
+                        .subscribe({}, {})
+                    reLoadSubscriptions()
+                } else {
+                    viewState.showToast(resourceProvider.getStringResource(R.string.unsubscribe_from_trader_alarm_message))
+                }
             }
+        }
+
+        private fun needDeleteSubscription(pos: Int): Boolean {
+            return if (subscriptionClickPos == pos) {
+                clearSubscriptionClickPos()
+                true
+            } else {
+                subscriptionClickPos = pos
+                false
+            }
+        }
+
+        private fun reLoadSubscriptions(delayMills: Long = 200) {
+            Handler(Looper.getMainLooper()).postDelayed(
+                {
+                    loadSubscriptions()
+                }, delayMills
+            )
         }
     }
 
-    private fun reLoadSubscriptions(delayMills: Long = 200) {
-        Handler(Looper.getMainLooper()).postDelayed(
-            {
-                loadSubscriptions()
-            }, delayMills
-        )
+    override fun attachView(view: TraderMeObservationView?) {
+        super.attachView(view)
+        clearSubscriptionClickPos()
+    }
+
+    private fun clearSubscriptionClickPos() {
+        subscriptionClickPos = -1
     }
 
     override fun onFirstViewAttach() {
